+++ util/settings/settings_test.go
@@ -1444,3 +1444,18 @@ allowedAudiences: ["aud1", "aud2"]`},
 		})
 	}
 }
+
+func TestReplaceStringSecret(t *testing.T) {
+	secretValues := map[string]string{"my-secret-key": "my-secret-value"}
+	result := ReplaceStringSecret("$my-secret-key", secretValues)
+	assert.Equal(t, "my-secret-value", result)
+
+	result = ReplaceStringSecret("$invalid-secret-key", secretValues)
+	assert.Equal(t, "$invalid-secret-key", result)
+
+	result = ReplaceStringSecret("", secretValues)
+	assert.Equal(t, "", result)
+
+	result = ReplaceStringSecret("my-value", secretValues)
+	assert.Equal(t, "my-value", result)
+}
